Add files via upload
This script is used to update company systems OS and corresponding applications that are on the system (depending on what is released for any given month). The company systems reside on an air-gapped network with no access to the internet. 

This script will call on other sub-scripts, however, those sub-scripts have yet been sanitized of proprietary labels and references. They will be uploaded at a later date.